+++ packages/sitecore-jss-angular/src/services/placeholder.token.ts
@@ -62,9 +62,6 @@ export const PLACEHOLDER_MISSING_COMPONENT_COMPONENT = new InjectionToken<Type<u
 export const PLACEHOLDER_HIDDEN_RENDERING_COMPONENT = new InjectionToken<Type<unknown>>(
   'Sc.placeholder.hiddenRenderingComponent'
 );
-export const PLACEHOLDER_JSS_STATE = new InjectionToken<ComponentNameAndType[]>(
-  'Sc.placeholder.state'
-);
 export const DYNAMIC_COMPONENT = new InjectionToken<Type<unknown> | { [s: string]: unknown }>(
   'Sc.placeholder.dynamicComponent'
 );